backend/requirements.txt
torch>=2.4.0
numpy==1.26.2
fastapi==0.104.1
uvicorn==0.24.0
transformers>=4.45.0
peft>=0.18.0
accelerate>=0.34.0
pydantic==2.5.2
python-dotenv==1.0.0
sympy==1.12
matplotlib==3.8.2
pytesseract==0.3.10
PyPDF2==3.0.1
pdf2image==1.16.3
SpeechRecognition==3.10.0
Pillow==10.1.0
requests==2.31.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-google-genai>=0.0.3
google-generativeai>=0.3.0
python-multipart